Add support for using the SPRAM of UP5K
We allocate 2 of the 4 32kbyte banks to create a 64kbyte RAM memory
region for the processor.

We could use all banks but 64k should be sufficient for most
applications. The other two banks remain available for other
purposes, like an instruction cache when we read our programs
from flash memory.